Trello #225 - "Object is shared with user" notification
Adds a function that injects a basic privatemsg form in any form passed to it
and implements hook_form_FORM_ID_alter() to inject this on the basic collection
share form. Automates the subject and message for this form for ease of use,
but still requires the user to input the recipients.